Remove Telecom dependency on LockTaskController
The LockTaskController contains logic to detect emergency call intent,
however this is an @hide constant that only works for AOSP, since
some OEMs use different emergency dialers.

Move this dep to the non-mainline Telephony dependency in order
to break the dependency on a hidden Telecom constant.

Bug: 445997472
Test: presubmit + atest WmTests
Flag: EXEMPT PURE_REFACTOR
Change-Id: Iba78670fcd2de51f4fae95e6b5c9c22fa4a59ab4