Use a color layer to reveal new content of rotation animation
If alpha < 1 is set on display layer, all visible children layers
are also affected, which may look flickering because the layer
which occludes other layers becomes translucent.

So this makes the display layer keep alpha 1 and use a layer with
the same color as background that animates from alpha 1 to 0, that
will have the same effect as fading the content in but it can avoid
the blending case of translucent layers.

Bug: 438615184
Flag: com.android.window.flags.no_alpha_rotation_enter_animation
Test: Trigger display rotation.
Change-Id: Id0b8b4452313ae118ac23a313165f949739aa985