Make sure outer-most class has some @Ravenwood annotation...
if any of its members has any.

This is not technically required, but improves readability.

This CL introduces the logic and tests, but still doesn't enable it in
the Ravenwood side, because there are still some violations
in other directories. We'll fix them as follow-up and
update ravenwood-standard-options.txt later to enable the check.

Bug: 460450157
Test: presubmit
Flag: TEST_ONLY

Change-Id: I95c4ab3845175d02eb62714c4bd687e52de85818