[flexiglass] Move the falsing check to the motionEvent handler
We check falsing after the motion event occurs, store
the falsing state (whether to allow or disallow the current
in-progress scene change), and then use this
falsing state in the eventual check for canChangeScene.

We intentionally update the transitioningTo and
isTransitionUserInputOngoing to start sharing
immediately (Eagerly) so their values are always
kept updated for when onMotionEvent gets called
and evaluates whether to check falsing for gesture
in the current transition (if isTransitionUserInputOngoing=true).

Note: This change is over-generous to touches that *may*
resolve to a contentKey change. Ideally, in the future, if
the transition is not going to be fulfilled, we should
not check FalsingManager.isFalseTouch because:
  1. There's no need to reject a content change if the content
  change won't be requested.
  2. The bigger issue: Checking the falsing manager attributes
  the touch as valid (ie: not a false touch). However, a transition
  that only partially occurs and doesn't get fulfilled *is* likely
  to be a false touch and therefore shouldn't not be attributed as
  valid.

Flag: com.android.systemui.scene_container
Bug: 461821809
Test: atest SceneContainerViewModelTest
Test: enable logs (adb shell cmd statusbar echo -b
SceneFrameworkLog:DEBUG, adb shell setprop log.tag.FalsingManager DEBUG,
adb shell setprop log.tag.FalsingCollector DEBUG), observe falsing check logs
Test: from lockscreen, swipe down entire screen for QS doesn't false
Test: Swipe down twice on lockscreen to quick settings. Press quick
settings edit button. Observe device doesn't false.
Test: Do a quick/short swipe up from lockscreen to bouncer,
observe gesture falses

Change-Id: I53a31c3a657e41698f11561d0e224f5ec69b024f