Restrict when bouncerOverlayNotificationAlpha emits
This bouncer expansion-based alpha was added to correctly fade notifications in and out when transitioning between bouncer and lockscreen, but the flow was also emitting during transitions from bouncer to anywhere else, which faded notifications in even when they shouldn't have.

This change limits the flow to only emit when transitioning between bouncer and lockscreen, and then 1 when bouncer is fully gone to leave the alpha in a correct state for HUNs. During transitions between bouncer and anywhere else, it now does not emit any values so that alphaForTransitions can either receive relevant values from any other transition flows or simply not change the notification alpha during that time.

Fixes: 452888658
Test: SharedNotificationContainerViewModelTest; manual by going from bouncer to all possible other scenes (AOD, shade, quick settings, gone, occluded), confirming that notifications no longer flicker on the bouncer->AOD transition, and are at least not worse on other transitions
Flag: com.android.systemui.scene_container
Change-Id: I44edb242e9da62de2e24d0a5696c480dce1deab4