No-op if shade is not open upon shade collapse
Currently when the shade collapses, we attempt to change to a home
scene regardless of the current scene state. This change makes it so
that we do not make a scene change request if the shade is not open.

This fixes an issue where returning from edit mode to glanceable hub
ends up incorreclty on the Lockscreen scene because of the shade
collapse.

Test: atest ShadeInteractorSceneContainerImplTest
Test: manually verified returning from edit mode to hub lands on hub
Fix: 460021897
Flag: com.android.systemui.scene_container
Change-Id: Ib284c6ef0df24a4c95fbbe1a7bda2bbc545d205e