Support InsightCollection in notification renderer
In order to support multiple notification actions to be added to a
single notifications, we need to process them together as a group and
make a single call to NMS to push the adjustments. This is because each
time you push new adjustments, it replaces the previous ones.

Fixes: 461553926
Test: atest NotificationActionRendererTest
Flag: android.service.personalcontext.enable_personal_context_service
Change-Id: Iff1b364d67ce5ad5395ebb1d5ce268892dbd8a01