Add missing early validation of fullscreen request
In Ia458d59126271f5bb7e5b68184a5f5c404607723,
validateMultiwindowFullscreenRequestLocked was modified to only perform
early validation when the delegation flag is enabled, but it was made to
only be called when shell transitions is disabled or in the pre-flag
path. This change adds the missing validation call when the flag is
enabled and the shell transition will be started.

Like before, a rejection means the transition should be aborted and the
result reported, but an approval only means it's ok to start the
transition for WMShell to actually approve/reject and report the result
to the client.

Flag: com.android.window.flags.delegate_request_fullscreen_handling_to_shell
Bug: 296268915
Test: test the request is reject if the window is not the top focused
task

Change-Id: Iaaf4c721ba984545ccb338a731d9134920e22795