Fix LockScreenMinimalismCoordinatorTest#topUnseenIdentifier
LockScreenMinimalismCoordinator only elevate a top unseen notification
on KEYGUARD, not even locked shade. topUnseenIdentifier test calls
setShadeAndQsExpansionThenWait(0f, 1f), which should set the
StatusBarState to SHADE_LOCKED, before assuming the
topOngoingNotificationKey. The test used to pass on top of a mistake
that setShadeAndQsExpansionThenWait(0f, 1f) didn't change the
StatusBarState.

However, with recent updates, the StatusBarState was updated at calling
setShadeAndQsExpansionThenWait and this test starts to fail.

This change fixes the test by make sure the Shade and QS are collapsed,
and the StatusBarState is KEYGUARD before trying to update the
topOngoingNotificationKey.

Fix: 462137409
Test: atest LockScreenMinimalismCoordinatorTest#topUnseenIdentifier
Flag: com.android.systemui.scene_container
Change-Id: I859293361a785b49fdad991af877d7d2942f7868