Add new PiP2 PipTransitionState to TvPipController
Refactor TvPipController to be a PipTransitionState.PipTransitionStateChangedListener such that it responds to state changes and no longer needs to manage state internally.
Update TvPipTransition to make state updates based on the stage of enter transition.
Add dagger switch between pip1 and pip2 TvPipController implementations.
Note: The TvPipController functionality is not complete. The next step would be to implement the TvPipScheduler, which the TvPipController will interact with.

Test: m
Flag: com.android.wm.shell.enable_pip2_on_tv
Bug: 403347147
Change-Id: I20fbb9b551c2c99e932793b7bd3516e548668e59