SettingsProvider: Reduce size of aconfig data
When aconfig was migrated out of system_server into a separate daemon,
code was added to compare flag values between the two sources. That
comparison has since been removed, but SettingState still retains a map
of all aconfig default flag state. That map is quite large - about 900KB
on cf_x86_64_phone-next.

Although the comparison code is gone, the map is now used in one place
to check whether a flag is present and in another place to check whether
a flag is read-only. As such, we can trim down the retained data to
exactly what is needed. Doing this reduces the retained size to ~225KB.
Around 1/3 of the savings comes from de-duplicating aconfig namespace
and package strings, and the other 2/3 comes from switching from a
HashMap of AconfigdFlagInfo to a ArrayMap of ArraySets of flag names.

This change also removes some code that was only involved in populating
unused portions of the AconfigdFlagInfo objects, which is now dead.

Bug: None
Test: Examine heap dumps
Flag: EXEMPT BUGFIX
Change-Id: Ide7dc49926947b9c371bc66d40753e83ca9eb061