Add API for notifying domain selection emergency mode
This change is to add a new DomainSelectionEmergencyModeListener API,
which allows listeners to monitor the emergency mode state when AP
domain selection is enabled.
This helps applications implement the necessary requirements based on
the emergency mode when an emergency call or SMS is being initiated.

Bug: 423761083
Flag: com.android.internal.telephony.flags.domain_selection_emergency_mode_notification
Test: atest CtsTelephonyTestCases:TelephonyCallbackTest
Test: atest CtsTelephonyTestCases:TelephonyRegistryManagerTest
Test: atest FrameworksTelephonyTests

Change-Id: I3b4c6e489927ab263b65af589a4d090116f04dd5