Add TransitionMixpatcher Base Structure
Introduces TransitionMixpatcher which will eventually
be the main interface between WM Transitions and Shell
Transition-animation capabilities. It will replace
most of the functionality currently in Transitions.java.

The main design goal of this new system is to support
the recent uptick in complicated interactions between
existing and newer windowing features. Specifically,
there are now many UX actions which move "ownership"
of containers between feature organizers or otherwise
involve multiple organizers' containers in one logical
transition.

In the current dispatch system, if a transition contains
containers that span multiple organizers, we have to
build bespoke logic for each of the combinations because
only one "Handler" gets to handle a transition. This
lead to a complicated "tree" of "MixedHandlers" which
attempt to break-apart a transition and sub-dispatch
to other handlers.

While this approach somewhat works, it isn't particularly
standardized which means we can't easily make
cross-cutting changes or features without touching
a bunch of systems which all work differently.

The Mixpatcher internalizes (and thus standardizes) this
"Mixing" requirement. Instead assigning a single handler
to a transition, Mixpatcher (itself) manages the
transition, breaks it into constituent animations, and
manages the lifecycle and communication with WM.

Instead of organizers implementing a Handler, they instead
participate in planning how a transition is distributed
among constituent animations and how containers are passed
from one organizer to another.

Centralizing the dispatch and abstracting the "animate a
set of containers" also allows us to solve a long-standing
request of supporting "universal" continuation/transfer
of ongoing animations into new transitions. In other words,
because Mixpatcher manages all the containers and animation
lifecycles, we can design the Animation APIs such that it
requires animators to use a start state and be able to
report the "current state" at any time. This way, each
animator implementation can remain independent while
providing enough information for Mixpatcher to smoothly
move containers between animations.

This change just introduces the class with enough
implementation to demonstrate/test the core usage
patterns.

Bug: 469422564
Test: TransitionMixpatcherTests
Flag: com.android.window.flags.transit_mixpatcher_base
Change-Id: I8b102240450ccedbd3787a40213cfa9727c2c8fc