Clear the lock task when the last running activity finishes
The lock-task was not removed immediately when finishing the
activity (due to package update). Since the app restarted
itself and start lock-task mode, the new task was added
to the same lock-task session.

The finishing task was removed later after the close transition
completed, which accidentally removed all lock tasks that are
in the same session.

Although we do clear the lock task when the last running activity
normally finishes, the current solution does not cover the case
where the activity was finished due to package update. So, moving
the check to #makeFinishingLocked to cover all possible paths.

Bug: 461662709
Test: manually verified by the setup flow
Flag: com.android.window.flags.clear_lock_task_when_task_end
Change-Id: I73073a9064675771f6678740b605f42470dc32e2