Updated goldens as reported in cf_phone in I28700010478358019
Presubmit Bypass: Bypassing since right now the test is running on multiple device targets and the values recorded for every device is different, which leads to failure.A fix for running the test on a specific device target is already in place (cl/840222846). In order to make it work, we need to update the golden values.


Test: QuickSettingsShadeTransitionTest
Flag: EXEMPT TEST_ONLY

Change-Id: I61ab0919356660cda3f9535b2e96c23cde464231