ComputerControl: integrate with AppInteractionService
Reports any activities the agent is launching successfully to the
AppInteractionService for surfacing to the user.

Note that "blocked" launches (i.e. ones that end up on the agent's display,
but that the agent can neither see nor interact with and instead has to delegate
to the user) are not reported.

A follow-up will obtain AppInteractionAttribution from the agent and pipe
that through as well.

Bug: 454891648
Test: atest ComputerControlSessionImplTest
Flag: android.companion.virtualdevice.flags.computer_control_access
Flag: android.app.appfunctions.flags.enable_app_interaction_api
Change-Id: I91d0037e2a377f345597d723c9c6c153cf48e8f1