Reuse the lookahead invalidation trigger used by smartspace cards
The date / weather smartspace elements can suffer from a similar bug
that was seen in smartspace cards. This instance causes truncation of
the horizontal smartspace date / weather line when transitioning. This
change refactors that fix to be reusable, and then applies it to the
other smartspace elements.

Bug: 436817577
Test: Manually checked centering animation
Flag: com.android.systemui.scene_container
Change-Id: Ic77e5769a9208fac8687c0cba46ec61d531f02fe