Handle device entered state during dreaming
This change modifies how the `isDeviceEnteredOnBackStack` flow
determines whether the device is entered, shifting the check from
relying on the "lockscreen to gone" event, to checking the state of the
bottom scene on the back stack.

It also handles replacing the bottom scene on the back stack when device
is locked while the current scene is "above" lockscreen or gone. For
example, device locks after a timeout while dreaming.

Test: atest DeviceEntryInteractorTest
Test: atest SceneContainerStartableTest
Test: atest SceneBackInteractorTest
Test: atest DreamStartableTest
Test: verified device entered state with and without a secure auth
      method, when going between lockscreen, bouncer, shade, and gone
Test: verified device entered state with a secure auth method, that
      device remains entered immediately after dream starts from home
      screen, and switches to not entered after the 5s lock timeout
Test: the same CUJ with unsecured auth method doesn't work and is to be
      addressed in a follow-up CL
Bug: 453749672
Flag: com.android.systemui.scene_container
Change-Id: I8ddb95bb3c0aea6e0ebfa6f7878db12eb44b3af3