Disable self-kill recovery for display add/remove events
There are a few problems with enabling self-kill
recovery for display disconnection/reconnection.

1.Multiple tasks can be reparented to another display, and as
the self-kill recovery is async, it can mess up z-order.
2. At the end of some automated E2E tests, they clean up external
displays and tasks. Removing a display causes the tasks on the
display to be repatented to another display, so the self-kill
recovery can be triggered  and some tasks may be left on
another display unexpectedly.

For these reasons, self-kill recovery is disabled for display
disconnection and reconnection for now.

Flag: com.android.window.flags.enable_auto_recovery_from_self_kill
Bug: 463781180
Test: AppCompatDisplayCompatTests
Change-Id: If6ccf764d2007a0ab4b8092ef482ee7d8cf5a033