[flexiglass] External bouncer scene container
Pulls the bouncer overlay "up" so it renders above the Flexiglass scene
container and the notifications view (NSSL). This should help with
problems the occur when the bouncer overlay fails to occlude
notifications.

The design is quite simple:
- There's a second STL only for the bouncer overlay (and the obligatory
  "gone" scene as there must always be a scene below the overlay)
- Places the second STL in its own View, after both the view that
  contains SceneContainer and the one that contains NSSL
- Uses the onTransitionStart callback on both STLs to communicate
  bouncer-related transitions back and forth:
  - The real/original STL delegates show/hide bouncer overlay
    transitions to the new STL
  - The new STL delegates only the hide bouncer overlay transition back
    to the real/original STL
  - The new STL is the hoisted kind, making it easy to instantiate
    outside its composable. Callbacks are used to send the delegated
transition from the original STL to the hoisted new STL
  - The new STL uses callbacks but those ultimately route the delegated
    hide overlay to the old STL using its pre-existing
SceneContainerRepository system

Bug: 457668950
Test: manually verified that dragging up on the lock screen to show the
bouncer transitions normally, the dragging down on the bouncer or using
the back gesture transitions away from the bouncer normally, and that
unlocking in the bouncer transitions away from it normally
Test: manually tested unlocking using face unlock while showing the
bouncer
Flag: com.android.systemui.scene_container

Change-Id: If0c8b2e911d18c205db537c6ac983240d5789db8