Fix ComplicationHostViewControllerTest
The previous implementation of collectFlow relied on the view's
attachment state to manage the lifecycle of the flow collection. This
caused flakiness in tests where the view attachment signal was not
always reliable.

This change switches to a collectFlow overload that uses a
CoroutineScope and returns a Job. The lifecycle of this Job is now
manually managed within onViewAttached and onViewDetached, ensuring that
the flow collection for layout updates is started and stopped reliably
with the view controller's lifecycle, independent of the view's
attachment state.

Fixes: 458417176
Test: atest --rerun-until-failure 100 SystemUITests:ComplicationHostViewControllerTest
Flag: EXEMPT BUGFIX
Change-Id: I88386e7b26eddb8e338a8375d3b04e482234c17b