Fix inconsistent flagging
The KeyguardService state monitor callback and
KeyguardStateCallbackStartable triggering the IKeyguardStateCallback is
flagged by the SceneContainerFlag while the prevention of the same in
the KeyguardViewMediator is gated by two flags. Updating the flagging to
be consistent to allow for proper functionality when the
SceneContainerFlag is disabled but the wm refactor flag is enabled.

Bug: 462109093
Test: manual
Flag: NONE bugfix
Change-Id: Ic72221b0830504da7b736464e15e2bae1acbe9d5