Generate self signed certificates for LSKF keys
In order to perform security checks on LKSF keys we did not use Android
keystore for key generation. However, we would like to insert the keys
into the Android keystore after the security checks have been performed
on the keys. In order to do that, the kkeys should be signed with a
certificate. This process is automatically done in Android keystore
while generating keys. As a result, this CL generates self-signed
certificates with the private and public keys and inserts them into
Android keystore

Bug: 462118325
Test: atest LskfResetTests
Flag: android.app.lskfreset.flags.enable_lskf_reset_manager
Change-Id: I138945c287417c2691bdd2a2cde16b67f770bf5d