Wrap nativeStartNative{Process,ChildZygote} in a try-catch block
Currently, a crash in Native Zygote brings down the system server. This
is because nativeStartNative{Process,ChildZygote} throws IOException
upon crash, which is not caught by AMS as opposed to RuntimeException.
This commit puts them in a try-catch block to re-throw them as
RuntimeException for proper error handling.

Test: system server does not crash upon specifying an invalid shared library for Native App Zygote
Flag: android.os.native_app_zygote
Bug: 461414744
Change-Id: I432c89ac1fb9173718c7ad9b5625f16a54c7f855