Add native support to MemoryLimiter
The MemoryLimiter class watches for changes in the procstate of an
application.  When the procstate requires a change in the maximum
memory allowed to an app, the class asks the native layer to apply the
new memory limit.

The native layer includes a thread that monitors a process's cgroup
events, waiting for events that signal violation of a memory limit.
The Java layer is notified when an event is detected.  At this time,
the Java layer emits a single log message on an event.

Flag: com.android.server.am.memory_limiter_enable
Bug: 461546255
Test: atest
 * FrameworksServicesTests:com.android.server.am
 * FrameworksCoreTests:ActivityTests
 * FrameworksMockingServicesTests:com.android.server.am
Change-Id: Id96c4eece5c5f473cd26fa78eaf50aac1587d9dd