Add IDesktopTaskListener#onTaskAppearingInDeskWithOverviewShowing
To notify Launcher about a task launching while Overview is showing,
so that Overview can close itself if needed.

This is needed because when Overview is in screenshot mode, it cannot
rely on the launcher activity's onStop as a signal that overview should
close because onStop won't get called if RecentsWindow is being used or
if launcher will remain resumed (such as when using
show_home_behind_desktop).

This is only needed when the recents transition isn't running. In
live-tile mode, overview closes itself from
RecentsAnimationListener#onTasksAppeared

Flag: com.android.window.flags.better_desk_deactivation_in_recents_transition
Bug: 410914723
Test: In CD, enter Overview from a desk, enter screenshot mode by
swiping the status bar down and up, click the "App Info" button of a
fullscreen tile - verify the task launches in desktop and overview
closes.

Change-Id: I6f557dbc60c40b025fffe559d2ad31446e049c7a