Apply rotation transform if the orientation source is not top
For example, when display is in portrait, move a task to front:
 B: Translucent activity orientation=unspecified
 A: Opaque activity orientation=landscape
B does not provide orientation, so the orientation source is A which
is landscape. Then when resuming B, it should be applied with
landscape transform. And when making A visible, it should share the
same transform from B, because the orientation of orientation source
should be the same as the orientation of the launching fixed
rotation app. And if the orientation of non-top visible activity is
unspecified, the activity either stay in current display orientation
or use the same orientation as the launching fixed rotation app which
will affect the display orientation.

Fix: 464307133
Flag: EXEMPT BUGFIX
Test: atest DisplayContentTests#testUnspecifiedOrientationTranslucentTop
Change-Id: Iae269ab4a52406e27b4ff416be0dcf11cb2ed710