Make sure that we only await bubble bar position in bubble bar mode
Updates to the device config or the bubbles mode could potentially
happen in any order, so make sure that we are only pending the
bubble bar position if we are in bubble bar mode.

Also whenever we update the device config in BubblePositioner,
reset the pending signal based on the mode. This part isn't necessary
but is reasonable because it better reflects the value of the
underlying field.

Bug: 457332269
Flag: EXEMPT BUGFIX
Test: atest BubblePositionerTest
Test: manual
       - verify that expanded view bounds are correct when rotating
         in both modes
Change-Id: Ia44e5a8dc376d8556ae207d39a0cf5adf41a8a1b