Correct isAssignableFrom usage in CameraMetadata.getKeys()
The goal of CameraMetadata.getKeys is to iterate over all fields in a
class that have type `keyClass`. For instance, it may scan
`CaptureRequest` for all fields of type `CaptureRequest$Key`.

Previously, CameraMetadata.getKeys was incorrectly calling
`field.getType().isAssignableFrom(keyClass)` when attempting to filter
for fields of type `keyClass`. This backwards logic caused the code to
incorrectly select fields that were either an instance of keyClass, or
a superclass of keyClass (e.g. Object).

This caused problems in Robolectric, where a synthetic public Object is
added to Android classes during instrumentation.

Invert the `field.getType().isAssignableFrom(keyClass)` to be
`keyClass.isAssignableFrom(field.getType())`. This will include fields
of type `keyClass` and its subclasses.

Bug: 460839981
Test: cts camera2 tests
Flag: EXEMPT BUGFIX

Change-Id: I207272db51679345dea62d41d160212c64f860a3