Fix IMMS user switching race condition
If we have two full users A and B, with personal profile of A as
userId 0, work profile of A as userId 10, and user B as userId 20,
there is a scenario in which IMMS will be locked out of the correct
full user, incorrectly disallowing startInput and showSoftInput calls.
This can happen if we are in userId 10, schedule a switch to 20, and
then get a startInput on 0 before we can finish the scheduled switch.

This fixes the issue by defining a "target user", loosely meaning the
userId we aim to be in. This is either the mCurrentImeUserId if there is
no pending user switch, or the userId of the pending user switch
otherwise. We then refine the startInput userId check to allow it if it
is the same as this target user, or a profile of this target user.

This aims to disallow startInput calls on any profile of the current
full user, while we have a pending user switch to a different full user,
while maintaing all the current requirements.

Flag: EXEMPT BUGFIX
Test: atest MultiUserMockImeTest
Bug: 462531364
Change-Id: I01006ab8ff4f6739e106ec3b3a153e1881335f0d