VRI#checkThreadCompat should not make an effort to repost work to the correct thread
Previously, in compatibility mode rather than throw an exception at the caller we would attempt to fix their threading problem for them by posting the work to the correct thread.
This is a risky behavior change.
It's also not the framework's responsibility to solve app bugs.

Simplify checkThreadCompat such that when targeting newer SDKs it acts like checkThread, and when targeting older SDKs it just logs.
Remove the code to support reposting work to the correct thread.

Bug: 464275874
Flag: EXEMPT BUGFIX
Test: ViewRootImplTest
Change-Id: I06fa0451a10f13cff80ef3d165bc04f70474c0a8