[TDR] Remove TODOs linked to issue 233177466
This CL was generated by an AI agent and verified by the CL owner before
sending for review. The TODOs were identified as stale by TODO Removal
Agent (http://go/dead-todo-removal-agent).
* Detailed reasoning can be found in the CL comments.
* How to review this CL: http://go/tdr-review-instructions
* To opt-out from future reviews: http://go/tdr-opt-out.

The associated bug crbug.com/233177466 was closed as 'OBSOLETE',
indicating that the planned refactoring to move the mMinWidth and
mMinHeight fields is no longer going to be implemented. The TODO comment
is therefore irrelevant and can be removed as noise.

Flag: EXEMPT DOCS_ONLY
Change-Id: I1f5797adbbcd046a671b8d047489917811e47a01
Bug: 233177466